+++ src/App.tsx
@@ -1,19 +1,15 @@
 import React from "react";
-import { SafeAreaView } from "react-native";
 import ApolloClient from "apollo-boost";
 import { ApolloProvider } from "react-apollo";
-import { Header, MovieList } from "./components";
+import { AppNavigator } from "./common/AppNavigator";
 
 const client = new ApolloClient({
-    uri: "https://kit-awbadxsigm.now.sh/graphql",
+    uri: "https://kit-omgqlyfdcg.now.sh/graphql",
 });
 
 const App = () => (
     <ApolloProvider client={client}>
-        <SafeAreaView style={{ flex: 1 }}>
-            <Header />
-            <MovieList />
-        </SafeAreaView>
+        <AppNavigator />
     </ApolloProvider> 
 );