+++ src/common/AppNavigator.ts
@@ -1,0 +1,17 @@
+import { MovieList, MovieDetail } from "../screens";
+import { createStackNavigator } from "react-navigation"; 
+
+export const Screens = {
+    MovieList: "MovieList",
+    MovieDetails: "MovieDetails",
+};
+
+export const AppNavigator = createStackNavigator(
+    {
+        MovieList: MovieList,
+        MovieDetails: MovieDetail,
+    },
+    {
+        initialRouteName: Screens.MovieList,
+    },
+);