+++ src/screens/index.ts
@@ -1,0 +1,2 @@
+export { default as MovieDetail } from "./MovieDetail";
+export { default as MovieList } from "./MovieList";