+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Add react-navigation and movie details screen
+- Added tests and tweaked Jest to play nicely with react-navigation
+- Allowed movie on MovieDetails via both props and navigation 
+- Updated GraphQL endpoint as result of new server deploy
+- Updated GraphQL model as result of new schema deploy
+- Updated folder structure
+
 ## GraphQL types automatic generation
 - Please install apollo `npm install -g apollo`
 - Run `yarn generate:graphql` and the models will be available at src/queries/models